--- SwiftKitCore/src/main/java/org/swift/swiftkitcore/SwiftHeapObject.java
@@ -1,22 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// This source file is part of the Swift.org open source project
-//
-// Copyright (c) 2024 Apple Inc. and the Swift.org project authors
-// Licensed under Apache License v2.0
-//
-// See LICENSE.txt for license information
-// See CONTRIBUTORS.txt for the list of Swift.org project authors
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-//===----------------------------------------------------------------------===//
-
-package org.swift.swiftkitcore;
-
-
-/**
- * Represents a wrapper around a Swift heap object, e.g. a {@code class} or an {@code actor}.
- */
-public interface SwiftHeapObject {
-}
