--- SwiftKitCore/src/main/java/org/swift/swiftkitcore/SwiftValue.java
@@ -1,21 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// This source file is part of the Swift.org open source project
-//
-// Copyright (c) 2024 Apple Inc. and the Swift.org project authors
-// Licensed under Apache License v2.0
-//
-// See LICENSE.txt for license information
-// See CONTRIBUTORS.txt for the list of Swift.org project authors
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-//===----------------------------------------------------------------------===//
-
-package org.swift.swiftkitcore;
-
-/**
- * Represent a wrapper around a Swift value object. e.g. {@code struct} or {@code enum}.
- */
-public interface SwiftValue {
-}
