--- SwiftKitFFM/src/main/java/org/swift/swiftkitffm/FFMSwiftInstance.java
@@ -1,59 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// This source file is part of the Swift.org open source project
-//
-// Copyright (c) 2024 Apple Inc. and the Swift.org project authors
-// Licensed under Apache License v2.0
-//
-// See LICENSE.txt for license information
-// See CONTRIBUTORS.txt for the list of Swift.org project authors
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-//===----------------------------------------------------------------------===//
-
-package org.swift.swiftkitffm;
-
-import org.swift.swiftkitcore.SwiftArena;
-import org.swift.swiftkitcore.SwiftInstance;
-import org.swift.swiftkitcore.SwiftInstanceCleanup;
-
-import java.lang.foreign.GroupLayout;
-import java.lang.foreign.MemorySegment;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-public abstract class FFMSwiftInstance extends SwiftInstance {
-    /**
-     * The pointer to the instance in memory. I.e. the {@code self} of the Swift object or value.
-     */
-    public final MemorySegment $memorySegment() {
-        return MemorySegment.ofAddress(this.pointer());
-    }
-
-    /**
-     * The Swift type metadata of this type.
-     */
-    public abstract SwiftAnyType $swiftType();
-
-    /**
-     * The designated constructor of any imported Swift types.
-     *
-     * @param segment the memory segment.
-     * @param arena the arena this object belongs to. When the arena goes out of scope, this value is destroyed.
-     */
-    protected FFMSwiftInstance(MemorySegment segment, AllocatingSwiftArena arena) {
-        super(segment.address(), arena);
-    }
-
-    @Override
-    public SwiftInstanceCleanup makeCleanupAction() {
-        var statusDestroyedFlag = $statusDestroyedFlag();
-        Runnable markAsDestroyed = () -> statusDestroyedFlag.set(true);
-
-        return new FFMSwiftInstanceCleanup(
-                $memorySegment(),
-                $swiftType(),
-                markAsDestroyed
-        );
-    }
-}
